separate logic from presentation for price block
